improve decoupling and some other small fixes
