+++ justfile
@@ -20,6 +20,7 @@ setup:
   docker compose run --rm django python manage.py assign_random_labels_to_colleagues
   docker compose run --rm django python manage.py createsuperuser --noinput
   docker compose run --rm django python manage.py ensure_initial_user
+  docker compose run --rm django python manage.py setup_initial_user
 
 # Generate full dummy data: sync organizations + create dummy colleagues/assignments/placements
 load-full-data:
