+++ wies/rijksauth/admin.py
@@ -1,10 +1,19 @@
 from django.contrib import admin
+from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
 
 from .models import User
 
 
 @admin.register(User)
-class UserAdmin(admin.ModelAdmin):
+class UserAdmin(BaseUserAdmin):
+    ordering = ["email"]
     list_display = ["email", "first_name", "last_name", "is_active", "date_joined"]
-    list_filter = ["is_active", "is_staff"]
+    list_filter = ["is_active", "is_staff", "groups"]
     search_fields = ["first_name", "last_name", "email"]
+    fieldsets = (
+        (None, {"fields": ("email", "password")}),
+        ("Personal info", {"fields": ("first_name", "last_name")}),
+        ("Permissions", {"fields": ("is_active", "is_staff", "is_superuser", "groups", "user_permissions")}),
+        ("Important dates", {"fields": ("last_login", "date_joined")}),
+    )
+    add_fieldsets = ((None, {"classes": ("wide",), "fields": ("email", "password1", "password2")}),)
